requirements-server.txt
chardet==3.0.4
certifi==2020.11.8
click==7.1.2
efficientnet-pytorch==0.7.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
MarkupSafe==1.1.1
nmslib==2.0.6
numpy==1.19.2
opencv-python==4.5.1.48
packaging==20.8
Pillow==8.1.0
psutil==5.8.0
pybind11==2.6.2
pyparsing==2.4.7
requests==2.24.0
scikit-learn==0.24.0
scipy==1.5.4
threadpoolctl==2.1.0
timm==0.3.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
